Cleaned up the workspace by migrating this test to Reconstruction tests
